# openapi/target/classes/openapi.yml
openapi: 3.0.3
info:
  title: API oPROsita
  description: API REST para gestión de grupos, usuarios, contenido y simulacros
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /simulacros:
    get:
      tags:
        - simulacros
      operationId: getSimulacros
      summary: Obtener simulacros
      responses:
        '200':
          description: Lista de simulacros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Simulacro'

    post:
      tags:
        - simulacros
      operationId: createSimulacro
      summary: Crear un nuevo simulacro (PDF + descripción)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                descripcion:
                  type: string
                grupoId:
                  type: integer
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Simulacro creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulacro'
  /simulacros/{id}:
    delete:
      tags:
        - simulacros
      operationId: deleteSimulacro
      summary: Eliminar un simulacro
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Simulacro eliminado correctamente
  /profesores/{profesorId}/grupos:
    get:
      tags:
        - profesores
        - grupos
      operationId: getGruposByProfesor
      summary: Obtener grupos de un profesor
      parameters:
        - in: path
          name: profesorId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de grupos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grupo'
...
  /grupos/{grupoId}/reuniones:
    get:
      tags:
        - reuniones
      operationId: getReunionesByGrupo
      summary: Obtener reuniones de un grupo
      parameters:
        - in: path
          name: grupoId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de reuniones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reunion'
    post:
      tags:
        - reuniones
      operationId: createReunion
      summary: Crear una reunión para un grupo
      parameters:
        - in: path
          name: grupoId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reunion'
      responses:
        '201':
          description: Reunión creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reunion'
  /reuniones/{id}:
    put:
      tags:
        - reuniones
      operationId: updateReunion
      summary: Actualizar una reunión
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reunion'
      responses:
        '200':
          description: Reunión actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reunion'

    delete:
      tags:
        - reuniones
      operationId: deleteReunion
      summary: Eliminar una reunión
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Reunión eliminada correctamente
...
components:
  schemas:
    Grupo:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        meses:
          type: array
          items:
            $ref: '#/components/schemas/Mes'
    Mes:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
...
    Simulacro:
      type: object
      properties:
        id:
          type: integer
        descripcion:
          type: string
        archivoId:
          type: integer
        grupoId:
          type: integer
...
    Reunion:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        descripcion:
          type: string
        fechaHora:
          type: string
          format: date-time
        enlace:
          type: string
        grupoId:
          type: integer
...