# openapi/target/generated-sources/openapi/src/main/java/com/oprosita/backend/api/ProfesoresApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.oprosita.backend.api;

import com.oprosita.backend.model.generated.Grupo;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-05-23T22:26:08.390066400+02:00[Europe/Madrid]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "profesores", description = "the profesores API")
public interface ProfesoresApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /profesores/{profesorId}/grupos : Obtener grupos de un profesor
     *
     * @param profesorId  (required)
     * @return Lista de grupos (status code 200)
     */
    @Operation(
        operationId = "getGruposByProfesor",
        summary = "Obtener grupos de un profesor",
        tags = { "profesores", "grupos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de grupos", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Grupo.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/profesores/{profesorId}/grupos",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Grupo>> getGruposByProfesor(
        @Parameter(name = "profesorId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("profesorId") Integer profesorId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"meses\" : [ { \"id\" : 6, \"nombre\" : \"nombre\" }, { \"id\" : 6, \"nombre\" : \"nombre\" } ], \"id\" : 0, \"nombre\" : \"nombre\" }, { \"meses\" : [ { \"id\" : 6, \"nombre\" : \"nombre\" }, { \"id\" : 6, \"nombre\" : \"nombre\" } ], \"id\" : 0, \"nombre\" : \"nombre\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}