# openapi/target/generated-sources/openapi/src/main/java/com/oprosita/backend/api/ReunionesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.oprosita.backend.api;

import com.oprosita.backend.model.generated.Reunion;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-05-23T22:26:08.390066400+02:00[Europe/Madrid]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "reuniones", description = "the reuniones API")
public interface ReunionesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /grupos/{grupoId}/reuniones : Crear una reunión para un grupo
     *
     * @param grupoId  (required)
     * @param reunion  (required)
     * @return Reunión creada (status code 201)
     */
    @Operation(
        operationId = "createReunion",
        summary = "Crear una reunión para un grupo",
        tags = { "reuniones" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Reunión creada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Reunion.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/grupos/{grupoId}/reuniones",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Reunion> createReunion(
        @Parameter(name = "grupoId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("grupoId") Integer grupoId,
        @Parameter(name = "Reunion", description = "", required = true) @Valid @RequestBody Reunion reunion
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"descripcion\" : \"descripcion\", \"fechaHora\" : \"2000-01-23T04:56:07.000+00:00\", \"titulo\" : \"titulo\", \"enlace\" : \"enlace\", \"id\" : 0, \"grupoId\" : 6 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /reuniones/{id} : Eliminar una reunión
     *
     * @param id  (required)
     * @return Reunión eliminada correctamente (status code 204)
     */
    @Operation(
        operationId = "deleteReunion",
        summary = "Eliminar una reunión",
        tags = { "reuniones" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Reunión eliminada correctamente")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/reuniones/{id}"
    )
    
    default ResponseEntity<Void> deleteReunion(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /grupos/{grupoId}/reuniones : Obtener reuniones de un grupo
     *
     * @param grupoId  (required)
     * @return Lista de reuniones (status code 200)
     */
    @Operation(
        operationId = "getReunionesByGrupo",
        summary = "Obtener reuniones de un grupo",
        tags = { "reuniones" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de reuniones", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Reunion.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/grupos/{grupoId}/reuniones",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Reunion>> getReunionesByGrupo(
        @Parameter(name = "grupoId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("grupoId") Integer grupoId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"descripcion\" : \"descripcion\", \"fechaHora\" : \"2000-01-23T04:56:07.000+00:00\", \"titulo\" : \"titulo\", \"enlace\" : \"enlace\", \"id\" : 0, \"grupoId\" : 6 }, { \"descripcion\" : \"descripcion\", \"fechaHora\" : \"2000-01-23T04:56:07.000+00:00\", \"titulo\" : \"titulo\", \"enlace\" : \"enlace\", \"id\" : 0, \"grupoId\" : 6 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /reuniones/{id} : Actualizar una reunión
     *
     * @param id  (required)
     * @param reunion  (required)
     * @return Reunión actualizada (status code 200)
     */
    @Operation(
        operationId = "updateReunion",
        summary = "Actualizar una reunión",
        tags = { "reuniones" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Reunión actualizada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Reunion.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/reuniones/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Reunion> updateReunion(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "Reunion", description = "", required = true) @Valid @RequestBody Reunion reunion
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"descripcion\" : \"descripcion\", \"fechaHora\" : \"2000-01-23T04:56:07.000+00:00\", \"titulo\" : \"titulo\", \"enlace\" : \"enlace\", \"id\" : 0, \"grupoId\" : 6 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}