# openapi/target/generated-sources/openapi/src/main/java/com/oprosita/backend/api/SimulacrosApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.oprosita.backend.api;

import com.oprosita.backend.model.generated.Simulacro;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-05-23T22:26:08.390066400+02:00[Europe/Madrid]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "simulacros", description = "Operaciones con simulacros de examen")
public interface SimulacrosApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /simulacros : Crear un nuevo simulacro (PDF + descripción)
     *
     * @param descripcion  (optional)
     * @param grupoId  (optional)
     * @param file  (optional)
     * @return Simulacro creado (status code 201)
     */
    @Operation(
        operationId = "createSimulacro",
        summary = "Crear un nuevo simulacro (PDF + descripción)",
        tags = { "simulacros" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Simulacro creado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Simulacro.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/simulacros",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    default ResponseEntity<Simulacro> createSimulacro(
        @Parameter(name = "descripcion", description = "") @Valid @RequestParam(value = "descripcion", required = false) String descripcion,
        @Parameter(name = "grupoId", description = "") @Valid @RequestParam(value = "grupoId", required = false) Integer grupoId,
        @Parameter(name = "file", description = "") @RequestPart(value = "file", required = false) MultipartFile file
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"descripcion\" : \"descripcion\", \"archivoId\" : 6, \"id\" : 0, \"grupoId\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /simulacros/{id} : Eliminar un simulacro
     *
     * @param id  (required)
     * @return Simulacro eliminado correctamente (status code 204)
     */
    @Operation(
        operationId = "deleteSimulacro",
        summary = "Eliminar un simulacro",
        tags = { "simulacros" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Simulacro eliminado correctamente")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/simulacros/{id}"
    )
    
    default ResponseEntity<Void> deleteSimulacro(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /simulacros : Obtener simulacros
     *
     * @return Lista de simulacros (status code 200)
     */
    @Operation(
        operationId = "getSimulacros",
        summary = "Obtener simulacros",
        tags = { "simulacros" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de simulacros", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Simulacro.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/simulacros",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Simulacro>> getSimulacros(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"descripcion\" : \"descripcion\", \"archivoId\" : 6, \"id\" : 0, \"grupoId\" : 1 }, { \"descripcion\" : \"descripcion\", \"archivoId\" : 6, \"id\" : 0, \"grupoId\" : 1 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}